+++ src/spool/base.py
@@ -63,9 +63,14 @@ class LOSArrow(pg.GraphicsWidget, pg.GraphicsWidgetAnchor):
             pen=(255, 255, 255),
             pxMode=True)
 
-        self.label = QtGui.QGraphicsSimpleTextItem(
-            'Towards Sat.', parent=self)
-        self.label.setBrush(pg.mkBrush(255, 255, 255, 180))
+        self.label = pg.LabelItem(
+            'Towards Sat.',
+            justify='right', size='8pt',
+            parent=self)
+        self.label.anchor(
+            itemPos=(1., -1.),
+            parentPos=(1., 0.))
+        # self.label.setBrush(pg.mkBrush(255, 255, 255, 180))
         # self.label.setFont(QtGui.QFont(
         #     "Helvetica", weight=QtGui.QFont.DemiBold))
 
@@ -96,9 +101,7 @@ class LOSArrow(pg.GraphicsWidget, pg.GraphicsWidgetAnchor):
         rect_label = self.label.boundingRect()
         rect_arr = self.arrow.boundingRect()
 
-        self.label.setPos(
-            0.,
-            rect_label.height()*1.33)
+        self.label.setPos(-rect_label.width()/2., rect_label.height()*1.33)
 
     def setParentItem(self, parent):
         pg.GraphicsWidget.setParentItem(self, parent)